backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
transformers==4.35.0
torch==2.1.0
pillow==10.1.0
python-multipart==0.0.6
celery==5.3.4
gunicorn==21.2.0
whitenoise==6.6.0
